backend/requirements.txt
Django==5.2
djangorestframework==3.15.2
django-cors-headers==4.4.0
PyMySQL==1.1.0
mysqlclient==2.2.1
Pillow==10.4.0
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
django-filter==24.2